Completely rewriting everything
I've decided to make this thing into a proper functioning loader, And since this project needs a major refactor, I've decided that doing a complete overhaul & re-writing from the scratch is a better approach.
On this commit, I've coded some boilerplate codes that will help me on implementing the loader, And I've added a support for pk3 files -- which means I can render with a proper texturing instead of plain lightmap-only rendering.